experiments/pyproject.toml
[project]
name = "experiments"
version = "1.0"
description = "Add your description here"
readme = "../README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "humanize>=4.15.0",
    "matplotlib>=3.10.8",
    "netifaces>=0.11.0",
    "pandas>=3.0.1",
    "protobuf==4.21.12",
    "pyyaml>=6.0.3",
]
